Add ArduinoOTA lib and update with motor delay variable
It seems the 1.0.0 and 1.0.1 versions of ArduinoOTA have changed their
interface enough to break fat-cat-fix. Including an older copy of the
library to maintain function.

Updating the EEPROM schema to include a new "delay" variable to set the
amount of time the motor will run to dispense a quarter cup. After
changing cat foods, the previous delay was too long and the motor would
dispense too much food. Now the delay can be set to between 1000 and
9999 milliseconds.